+++ packages/node-sdk/src/index.ts
@@ -12,4 +12,5 @@ export * from "./session.js";
 export * from "./store.js";
 export * from "./ui.js";
 export * from "./wire.js";
+export * from "./android.js";
 export * from "./app.js";